12/6 update

+++ Nova/src/app/components/navbar/navbar.component.ts
@@ -7,6 +7,7 @@ import { Product } from 'src/app/interfaces/product';
 import { DataService } from 'src/app/services/data.service';
 import { Subscription } from 'rxjs';
 import { Router } from '@angular/router';
+import {AuthService} from "../../services/auth.service";
 
 
 @Component({
@@ -21,14 +22,13 @@ export class NavbarComponent implements OnInit, OnDestroy {
   productNames: String[] = [];
   productsService: ProductsService;
   searchForm!: FormGroup;
-
-  username: String |null = sessionStorage.getItem("username");
-
+  username: String | null = sessionStorage.getItem("username");
   message!: String;
   sent!: String;
   subscription!: Subscription;
-//   search: String = '';
+  test: String = '#Words, :)! ##More Words! ###.MD?';
   navbarOpen = false;
+
 
   ngOnInit(): void {
     this.productsService.getProducts().subscribe(data => {
@@ -43,6 +43,9 @@ export class NavbarComponent implements OnInit, OnDestroy {
       }
       this.subscription = this.data.currentMessage.subscribe(message => this.message = message)
       this.subscription = this.data.sentStatus.subscribe(sent => this.sent = sent)
+      // console.log('Before: ', this.test);
+      // this.test = this.test.replace(/#/g, '');
+      // console.log('After: ', this.test);
     })
   }
 
@@ -53,19 +56,21 @@ export class NavbarComponent implements OnInit, OnDestroy {
       this.showDropDown = false;
     }, 100)
     //console.log('...timeout passed.');
-    
+
   }
 
   //Check If the value of the form group exists or not
   toggleSearchOn() {
-    if (this.searchForm.value.search === null || this.searchForm.value.search === '') {
-      this.showDropDown = false;
-    } else {
-      this.showDropDown = true;
-    }
+    this.showDropDown = !(this.searchForm.value.search === null || this.searchForm.value.search === '');
   }
 
-  constructor( private fb: FormBuilder, _productsService: ProductsService, private data: DataService, private router: Router) {
+
+  constructor( private fb: FormBuilder,
+               private _productsService: ProductsService,
+               private data: DataService,
+               private router: Router,
+               private auth: AuthService) {
+
     this.initForm()
     this.productsService = _productsService;
   }
@@ -87,11 +92,11 @@ export class NavbarComponent implements OnInit, OnDestroy {
     let a = this.searchForm.value.search;
     this.data.changeMessage(a);
     this.data.changeSent('true');
-    
+
   }
 
   selectValue(value: any) {
-    
+
     this.searchForm.patchValue({"search": value});
     this.showDropDown = false;
     this.searchFor({'search': value})
@@ -111,13 +116,18 @@ export class NavbarComponent implements OnInit, OnDestroy {
   toggleNavbar() {
     this.navbarOpen = !this.navbarOpen
     console.log("clicked")
-  }  
+  }
 
   logout(){
-    sessionStorage.clear();
-    // alert("You are now logged out")
-    console.log("logged out")
-    // this.router.navigate(['/'])
+    this.auth.logout()
+      .subscribe(resp => {
+        if (resp.body == 'Successful Logout')
+        {
+          sessionStorage.clear();
+          console.log("logged out")
+          // this.router.navigate(['/'])
+        }
+      })
   }
 }
